+++ src/main/java/lasersharks/FishBot.java
@@ -6,7 +6,7 @@ package lasersharks;
  * @author Sytze
  *
  */
-public class FishBot implements Fish {
+public class FishBot extends Fish {
 
   /**
    * This value is used to modify the speed of the fishes that are generated. The generated speed is
@@ -25,74 +25,17 @@ public class FishBot implements Fish {
    */
   private static int chanceForLeft = 50;
 
-  private Position position;
-  private float size;
-  private int speed;
-  private Direction direction;
-
   /**
-   * The constructor for fishbot.
-   * @param pos the position where the fishbot is.
-   * @param siz the size of the fishbot.
-   * @param sp the speed parameter of the fishbot.
-   * @param dir the direction the fishbot is heading to.
+   * Constructor class for FishBot.
+   * @param position initial position
+   * @param size init size
+   * @param speed init speed
+   * @param direction init direction
    */
-  public FishBot(Position pos, float siz, int sp, Direction dir) {
-    position = pos;
-    size = siz;
-    speed = sp;
-    direction = dir;
+  public FishBot(Position position, float size, int speed, Direction direction) {
+    super(position, size, speed, direction);
   }
-
-  /**
-   * @return the position
-   */
-  @Override
-  public Position getPosition() {
-    return position;
-  }
-
-  /**
-   * @return the size
-   */
-  @Override
-  public float getSize() {
-    return size;
-  }
-
-  /**
-   * @return the size
-   */
-  public float getSpeed() {
-    return speed;
-  }
-
-  /**
-   * @return the direction
-   */
-  public Direction getDirection() {
-    return direction;
-  }
-
-  /**
-   * We calculate the distance between the fishes. The sum of the size of both fishes is our hitbox.
-   * Hitbox is now a circle, with size the radius in pixels.
-   * 
-   * @param fish we want to check if the fishbot collides with this fish,
-   * @return true if the fishes collide and false if not.
-   */
-  public boolean collision(Fish fish) {
-    return (position.calculateDistance(fish.getPosition()) < size + fish.getSize());
-  }
-
-  /**
-   * moves the fishbot using the direction and speed values of this object.
-   */
-  @Override
-  public boolean move() {
-    return position.updatePosition(direction, speed);
-  }
-
+  
   /**
    * This function creates a new FishBot with random values. This should be used to spawn fishes.
    * Starts on either the left side on
@@ -116,20 +59,5 @@ public class FishBot implements Fish {
     return new FishBot(new Position(posX, (int) (Position.getHeightPanel() * Math.random())),
         (int) Math.round(Math.random() * sizeModifier),
         (int) Math.round(Math.random() * speedModifier), dir);
-  }
-
-  /**
-   * this function checks if the fish is on the screen or not.
-   * 
-   * @return true if the fish is on the screen, and false if not.
-   */
-  public final boolean isOnScreen() {
-    return position.onScreen();
-  }
-
-  @Override
-  public String toString() {
-    return "FishBot [" + "position=" + position.toString() + ", size=" + size + ", speed=" + speed
-        + ", direction=" + direction + "]";
   }
 }
