+++ src/main/java/lasersharks/LaserShark.java
@@ -4,99 +4,15 @@ package lasersharks;
  * LaserShark class.
  *
  */
-public class LaserShark implements Fish {
-
-  private Position position;
-  private Direction direction;
-  private float size;
-
+public class LaserShark extends Fish {
   /**
-   * Initializes the laserShark(player).
-   * 
-   * @param position
-   *          the starting position
-   * @param direction
-   *          the starting direction
-   * @param size
-   *          the starting size
+   * Constructor class for FishBot.
+   * @param position initial position
+   * @param size init size
+   * @param speed init speed
+   * @param direction init direction
    */
-  public LaserShark(Position position, Direction direction, float size) {
-    this.position = position;
-    this.direction = direction;
-    this.size = size;
+  public LaserShark(Position position, float size, int speed, Direction direction) {
+    super(position, size, speed, direction);
   }
-
-  /**
-   * 
-   * @return the direction.
-   */
-  public Direction getDirection() {
-    return direction;
-  }
-
-  /**
-   * 
-   * @param direction the direction to set.
-   */
-  public void setDirection(Direction direction) {
-    this.direction = direction;
-  }
-
-  /**
-   * 
-   * @param position the position to set.
-   */
-  public void setPosition(Position position) {
-    this.position = position;
-  }
-
-  /**
-   * 
-   * @param size the size to set.
-   */
-  public void setSize(float size) {
-    this.size = size;
-  }
-
-  /**
-   * @return the position
-   */
-  @Override
-  public Position getPosition() {
-    return this.position;
-  }
-
-  /**
-   * 
-   * @param newDirection the direction to set to.
-   */
-  public void setNextMove(Direction newDirection) {
-    setDirection(newDirection);
-  }
-
-  /**
-   * @return the size.
-   */
-  @Override
-  public float getSize() {
-    return size;
-  }
-
-  /**
-   * 
-   * @param fish the fish we want to check if it collides with.
-   * @return true if it collides.
-   */
-  public boolean collision(Fish fish) {
-    return false;
-  }
-
-  /**
-   * moves the shark.
-   */
-  @Override
-  public boolean move() {
-    return getPosition().updatePosition(getDirection());
-  }
-
 }
