+++ src/test/java/lasersharks/FishTest.java
@@ -1,0 +1,37 @@
+/**
+ * 
+ */
+package lasersharks;
+
+import static org.junit.Assert.*;
+
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Test;
+
+/**
+ * @author Youri
+ *
+ */
+public class FishTest {
+
+  /**
+   * @throws java.lang.Exception
+   */
+  @Before
+  public void setUp() throws Exception {
+  }
+
+  /**
+   * @throws java.lang.Exception
+   */
+  @After
+  public void tearDown() throws Exception {
+  }
+
+  @Test
+  public void test() {
+    fail("Not yet implemented");
+  }
+
+}
